t01-dragon-strategy/core/screener.py
#!/usr/bin/env python3
"""
选股筛选器
整合API数据获取、策略筛选、结果输出
"""

import yaml
import logging
from typing import Dict, Any
import pandas as pd

from api_client import APIClient
from strategy import DragonHeadStrategy
from output_formatter import OutputFormatter

logger = logging.getLogger(__name__)


class StockScreener:
    """选股筛选器主类"""
    
    def __init__(self, config_path: str = "config.yaml"):
        with open(config_path, 'r', encoding='utf-8') as f:
            self.config = yaml.safe_load(f)
        
        # 初始化各模块
        self.api_client = APIClient(config_path)
        self.strategy = DragonHeadStrategy(self.config)
        self.output_formatter = OutputFormatter(self.config)
        
        logger.info("选股筛选器初始化完成")
    
    def fetch_market_data(self) -> Dict[str, Any]:
        """获取市场数据"""
        # 待实现：根据API文档获取所需数据
        # 目前返回示例数据
        logger.warning("数据获取功能待实现，等待API文档")
        
        # 示例数据结构
        market_data = {
            'stock_prices': pd.DataFrame(),  # 股票价格数据
            'stock_fundamentals': pd.DataFrame(),  # 基本面数据
            'sector_data': pd.DataFrame(),  # 板块数据
            'market_indices': pd.DataFrame()  # 大盘指数
        }
        
        return market_data
    
    def run_screening(self) -> Dict[str, Any]:
        """
        执行选股筛选流程
        返回：包含候选股票、排名、告警等信息的字典
        """
        logger.info("开始执行龙头战法选股筛选...")
        
        # 1. 获取数据
        market_data = self.fetch_market_data()
        
        # 2. 应用策略生成信号
        signals = self.strategy.generate_signals(market_data)
        
        # 3. 格式化输出
        output = self.output_formatter.format_output(signals)
        
        logger.info(f"选股完成，找到 {len(signals.get('candidates', []))} 个候选股票")
        
        return output
    
    def run_with_alert(self):
        """运行筛选并触发告警（如有）"""
        output = self.run_screening()
        
        # 检查告警条件
        alerts = output.get('alerts', [])
        if alerts:
            logger.warning(f"发现 {len(alerts)} 个告警")
            for alert in alerts:
                logger.warning(f"告警: {alert}")
        
        return output


if __name__ == "__main__":
    # 配置日志
    logging.basicConfig(
        level=logging.INFO,
        format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
    )
    
    # 运行筛选器
    screener = StockScreener()
    results = screener.run_screening()
    
    print("选股结果结构:", results.keys() if isinstance(results, dict) else "N/A")