t01-dragon-strategy/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
tushare>=1.2.0
requests>=2.28.0
pyyaml>=6.0
schedule>=1.1.0
